rebuilt the "make visible" part of the code. created new methods to make specific panels (it used to be imbeded when the data was built, but this created a problem when a data model has to be built as part of the preparation of another screen. keeping track of which screens will be visible after data was retrieved became difficult. screens are now made visible depending on the menu option selected (i.e. make the screen visible that is required by the menu option).
Also redesigned the logic of when data models are built to ensure data
is prepared to for all the menu options in a particular screen.

for example "showing all tables" in the explorer screen did not also
rebuilt the chart dataset. which means even though all tables where showing
in the explorer window, only the tables in the original query was available
for the chart.

also improved the used of the "currentlyinview" variable to keep track of which
swing panels are currently displayed in the card view.