Built the Table Column Chart.
Spent a long time getting the TextArea error message for no data charts
to update dynamically depending on the type of chart that is running.

Spent time getting the logic of the menu right so that the right chart display 
from the menu option. setvisible = false had to be called for previously created
charts