COLUMN NAMES Added for CSV select.
mySQL select generated to export data to CSV for dataTable BUT
still have to generate column names and actual script that exports the
data. 
also problem with delimeter and escape character.


Future improvement: allow user to run column analysys charts for only a
selection of columns.

Changed:  do not reset variables for columns selected and data filters
when the same table name is reselected from the data navigator window

.